==> scripts/02_HAARVIVAC_10X_S1_ADT_QC.Rmd <==
---
title: "HAARVIVAC 10X S1 ADT QC and Processing"
author: "Jolie Phan"
output: 
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: 4
date: "version `r format(Sys.time(), '%B %d, %Y')`"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries

```{r}
library(dplyr)
library(Seurat)
library(patchwork)
library(ggplot2)
library(ggpubr)
library(stringr)
library(RColorBrewer)
```

# Load data

```{r}
s1 <- readRDS(file = here::here("processed_data/20211118_s1_mdm_singlets.rds"))
save_output <- TRUE
```

# Gate ADT data

Gate cells by setting a cutoff expression level for each IgG isotype.

First, assign the markers to their respective IgG isotype control.

```{r}
# List protein surface markers
DefaultAssay(s1) <- "ADT"
rownames(s1)

iso_control <- c("Mouse-IgG1-Isotype-Control", "Mouse-IgG2a-Isotype-Control", "Mouse-IgG2b-Isotype-Control")

igg1 <- c("Mouse-IgG1-Isotype-Control", "CD3", "CD4", "CD8a", "GD", "TRAV1-2", 
          "CD62L", "CD28", "CD127", "CD95", "CXCR3", 
          "CXCR5", "CCR4", "CD25", "CD38", "CD161", 
          "CD103", "CD16", "CD56", "CD11b", "CD169", 
          "CD19", "CD123", "CD15", "CD163")

igg2a <- c("Mouse-IgG2a-Isotype-Control", "CCR7", "CD26", "HLA-DR", "CD14")

igg2b <- c("Mouse-IgG2b-Isotype-Control", "CD45RA", "CCR5", "CCR6", "CD11c", 
           "PD-L1", "CD20", "CD86")
```

Pull out expression level data for all samples and find the 99th percentile expression level for the IgG isotype controls.

```{r}
igg1_dat <- GetAssayData(s1, slot = "data", assay = "ADT") %>%
  as.data.frame() %>%
  tibble::rownames_to_column("Marker") %>%
  filter(Marker %in% igg1) %>%
  tidyr::pivot_longer(-Marker, names_to = "Cell", 
                      values_to = "Expression Level")
igg1_dat$Marker <- factor(igg1_dat$Marker, levels = igg1)

igg2a_dat <- GetAssayData(s1, slot = "data", assay = "ADT") %>%
  as.data.frame() %>%
  tibble::rownames_to_column("Marker") %>%
  filter(Marker %in% igg2a) %>%
  tidyr::pivot_longer(-Marker, names_to = "Cell", 
                      values_to = "Expression Level")
igg2a_dat$Marker <- factor(igg2a_dat$Marker, levels = igg2a)

igg2b_dat <- GetAssayData(s1, slot = "data", assay = "ADT") %>%
  as.data.frame() %>%
  tibble::rownames_to_column("Marker") %>%
  filter(Marker %in% igg2b) %>%
  tidyr::pivot_longer(-Marker, names_to = "Cell", 
                      values_to = "Expression Level")
igg2b_dat$Marker <- factor(igg2b_dat$Marker, levels = igg2b)

iso_control_dat <- GetAssayData(s1, slot = "data", assay = "ADT") %>%
  as.data.frame() %>%
  tibble::rownames_to_column("Marker") %>%
  filter(Marker %in% iso_control) %>%
  tidyr::pivot_longer(-Marker, names_to = "Cell", 
                      values_to = "Expression Level") %>%
  tidyr::pivot_wider(id_cols = Cell, names_from = "Marker",
                     values_from = "Expression Level")

# Find the 99th percentile expression level for each isotype control
igg1_percent <- quantile(iso_control_dat$"Mouse-IgG1-Isotype-Control", probs = 0.99)
igg2a_percent <- quantile(iso_control_dat$"Mouse-IgG2a-Isotype-Control", probs = 0.99)
igg2b_percent <- quantile(iso_control_dat$"Mouse-IgG2b-Isotype-Control", probs = 0.99)

igg1_percent
igg2a_percent
igg2b_percent
```

Visualize the ADT expression levels for all samples.

```{r, fig.width=12, fig.height=10}
igg1_xlab <- str_replace(igg1, "Mouse-IgG1-Isotype-Control", "Control")
igg2a_xlab <- str_replace(igg2a, "Mouse-IgG2a-Isotype-Control", "Control")
igg2b_xlab <- str_replace(igg2b, "Mouse-IgG2b-Isotype-Control", "Control")

p1 <- ggplot(igg1_dat, aes(x = Marker, y = `Expression Level`)) +
  geom_violin(scale = "width") +
  #geom_jitter(size = 0.5, shape = ".") + 
  ylim(0, 9) +
  ggtitle("Mouse IgG1 Isotype") +
  theme(plot.title = element_text(hjust = 0.5),
        axis.text.x=element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(labels = igg1_xlab) +
  geom_hline(yintercept = igg1_percent, linetype="dashed", color = "red")

p2 <- ggplot(igg2a_dat, aes(x = Marker, y = `Expression Level`)) +
  geom_violin(scale = "width") +
  #geom_jitter(size = 0.5, shape = ".") +
  ylim(0, 9) +
  ggtitle("Mouse IgG2a Isotype") +
  theme(plot.title = element_text(hjust = 0.5),
        axis.text.x=element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(labels = igg2a_xlab) +
  geom_hline(yintercept = igg2a_percent, linetype = "dashed", color = "red")

p3 <- ggplot(igg2b_dat, aes(x = Marker, y = `Expression Level`)) +
  geom_violin(scale = "width") +
  #geom_jitter(size = 0.5, shape = ".") +
  ylim(0, 9) +
  ggtitle("Mouse IgG2b Isotype") +
  theme(plot.title = element_text(hjust = 0.5),
        axis.text.x=element_text(angle = 45, hjust = 1)) +
  scale_x_discrete(labels = igg2b_xlab) +
  geom_hline(yintercept = igg2b_percent, linetype = "dashed", color = "red")

p1/(p2|p3)
```

**Spike 1** 99th percentile expression level for each IgG isotype control:

-   IgG1: 0.5702312

-   IgG2a: 1.00173

-   IgG2b: 0.4064592

Background subtract and set all negative expression values to 0.

```{r}
igg1_dat <- igg1_dat %>%
  mutate(bc_expression = `Expression Level` - igg1_percent) %>%
  tidyr::pivot_wider(id_cols = Marker, names_from = Cell, 
                     values_from = bc_expression) %>%
  tibble::column_to_rownames("Marker") 
igg1_dat[igg1_dat < 0] <- 0

igg2a_dat <- igg2a_dat %>%
  mutate(bc_expression = `Expression Level` - igg2a_percent) %>%
  tidyr::pivot_wider(id_cols = Marker, names_from = Cell, 
                     values_from = bc_expression) %>%
  tibble::column_to_rownames("Marker") 
igg2a_dat[igg2a_dat < 0] <- 0

igg2b_dat <- igg2b_dat %>%
  mutate(bc_expression = `Expression Level` - igg2b_percent) %>%
  tidyr::pivot_wider(id_cols = Marker, names_from = Cell, 
                     values_from = bc_expression) %>%
  tibble::column_to_rownames("Marker") 
igg2b_dat[igg2b_dat < 0] <- 0

bc_adt_dat <- igg1_dat %>%
  bind_rows(igg2a_dat) %>%
  bind_rows(igg2b_dat) %>%
  as.matrix()
```

Create a new ADT assay for the background-corrected expression levels.

```{r}
adt_assay <- GetAssay(s1, assay = "ADT")
Key(object = adt_assay) <- "adtbc_"
s1[["ADT_bc"]] <- adt_assay
s1[["ADT_bc"]] <- SetAssayData(s1[["ADT_bc"]], slot = "data", new.data = bc_adt_dat)
```

# Clustering and Annotation

The data is already normalized, so just proceed with the rest of the dimensionality reduction and clustering steps.

## Cluster on the ADT data

```{r, fig.width=6, fig.height=4}
DefaultAssay(s1) <- "ADT_bc"
VariableFeatures(s1) <- rownames(s1[["ADT_bc"]])
s1 <- ScaleData(s1) 
s1 <- RunPCA(s1, reduction.name = "apca", 
                         reduction.key = "APC_")
ElbowPlot(s1)

# Since there are only 35 markers, I'm going to use a standard euclidean distance
# matrix for visualization and clustering instead of PCA. 
adt_data <- GetAssayData(s1, slot = "data", assay = "ADT_bc")
adt_dist <- dist(t(adt_data))
s1[["umap_adt"]] <- RunUMAP(adt_dist, assay = "ADT_bc", reduction.key = "adtUMAP_")
s1[["adt_snn"]] <- FindNeighbors(adt_dist)$snn
s1 <- FindClusters(s1, resolution = 0.4, 
                               graph.name = "adt_snn")

DimPlot(s1, reduction = "umap_adt", label = TRUE, pt.size = 0.7) +
  ggtitle("Spike 1") +
  theme(plot.title = element_text(hjust = 0.5))
```

## Violin plots

Visualize surface protein expression and determine what the broad cell types are.

```{r}
lin_adt_feat <- c("CD3", "CD4", "CD8a", "GD", "TRAV1-2")
lin2_adt_feat <- c("CD14", "CD56", "CD19", "CD123", "CD15", "CD20", "CD163")
lin3_adt_feat <- c("CD16", "CD103", "CD161", "CD169", "CD11b", "CD11c")
act_adt_feat <- c("CD95", "CD25", "CD38", "CD26", "HLA-DR")
mem_adt_feat <- c("CD62L", "CD127", "CCR7", "CD45RA", "CD28")
ck_adt_feat <- c("CXCR3", "CXCR5", "CCR4", "CCR5", "CCR6")
fxn_adt_feat <- c("PD-L1", "CD86")
```

```{r, fig.width=12, fig.height=10}
lin_plot <- VlnPlot(s1, assay = "ADT_bc", features = lin_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

lin2_plot <- VlnPlot(s1, assay = "ADT_bc", features = lin2_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

lin3_plot <- VlnPlot(s1, assay = "ADT_bc", features = lin3_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

act_plot <- VlnPlot(s1, assay = "ADT_bc", features = act_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

mem_plot <- VlnPlot(s1, assay = "ADT_bc", features = mem_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

ck_plot <- VlnPlot(s1, assay = "ADT_bc", features = ck_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

fxn_plot <- VlnPlot(s1, assay = "ADT_bc", features = fxn_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

lin_plot
lin2_plot
lin3_plot
act_plot
mem_plot
ck_plot
fxn_plot

if(save_output) {
  cairo_pdf(file=here::here("out/20220323_S1_ADT_Violin_Plots.pdf"), width=12, height=10, onefile = TRUE)

  print(lin_plot)
  print(lin2_plot)
  print(lin3_plot)
  print(act_plot)
  print(mem_plot)
  print(ck_plot)
  print(fxn_plot)

  dev.off()
}
```

## Assign labels

Based on the ADT expression levels, these were the broad cell types assigned to each cluster.

Clusters 0, 1, 4, 5, 9, 10: CD4 T cells

Clusters 3, 6, and 8: CD8 T cells

Cluster 2: NK cells

Cluster 7: DCs

Cluster 11: B cells

```{r}
# Assign cell type labels
i1 <- grepl("0|1|4|5|9|10", s1@meta.data$adt_snn_res.0.4)
i2 <- grepl("3|6|8", s1@meta.data$adt_snn_res.0.4)
i3 <- grepl("2", s1@meta.data$adt_snn_res.0.4) 
i4 <- grepl("7", s1@meta.data$adt_snn_res.0.4) 
i5 <- grepl("11", s1@meta.data$adt_snn_res.0.4) 
s1@meta.data$adt_clusters <- NULL
s1@meta.data$adt_clusters[i1] <- "CD4 T cells"
s1@meta.data$adt_clusters[i2] <- "CD8 T cells"
s1@meta.data$adt_clusters[i3] <- "NK cells"
s1@meta.data$adt_clusters[i4] <- "DCs"
s1@meta.data$adt_clusters[i5] <- "B cells"

# Set colors for plotting
brewer.pal(5, "Set2")

s1_adt_colors <- c("CD4 T cells" = "#66C2A5",
                   "CD8 T cells" = "#FC8D62",
                   "NK cells" = "#E78AC3",
                   "DCs" = "#A6D854",
                   "B cells" = "#FFD92F")
```

## UMAP

```{r, fig.width=6, fig.height=4}
Idents(s1) <- "adt_clusters"
p <- DimPlot(s1, reduction = "umap_adt", label = FALSE, pt.size = 0.7, cols = s1_adt_colors) +
  ggtitle("Spike 1") +
  theme(plot.title = element_text(hjust = 0.5))

p$data$ident <- factor(x = p$data$ident, levels = c("CD4 T cells", "CD8 T cells", "NK cells", "DCs", "B cells"))

p
```

## Heat map

```{r, fig.width=12, fig.height=6}
adt_markers <- c(lin_adt_feat, lin2_adt_feat, lin3_adt_feat, act_adt_feat, mem_adt_feat, ck_adt_feat, fxn_adt_feat)
DoHeatmap(s1, features = adt_markers, label = FALSE, assay ="ADT_bc", size = 4) +
  ggtitle("Spike 1") +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_fill_gradientn(colors = brewer.pal(n = 10, name = "RdYlBu"))
```

## Dot plot

```{r, fig.width=8, fig.height=4}
Idents(s1) <- "adt_clusters"
levels(s1) <- c("B cells", "DCs", "NK cells", "CD8 T cells", "CD4 T cells")
adt_subset <- c("CD3", "CD4", "CD8a", "CD56")
adt_myeloid <- c("CD123", "CD16", "CD11b", "CD11c")
adt_bcell <- c("CD19", "CD20")
adt_markers <- c(adt_subset, adt_myeloid, adt_bcell)

# Make a Seurat dot plot just to easily pull the necessary data
s1_dotplot <- DotPlot(s1, features = adt_markers, assay ="ADT_bc") + 
  RotatedAxis() +
  ggtitle("S1") +
  theme(plot.title = element_text(hjust = 0.5))

# Pull the necessary data from the Seurat-generated dotplot 
s1_dotplot_data <- s1_dotplot$data

# Get count and proportion of each cell type
clust_count <- s1@meta.data %>%
  count(adt_clusters) %>%
  mutate(Freq = (n/sum(n))*100)

clust_count$adt_clusters <- factor(clust_count$adt_clusters, levels = c("CD4 T cells", "CD8 T cells", "NK cells", "DCs", "B cells"))

# Add cell count to each cluster name
s1_dotplot_data <- s1_dotplot_data %>%
  mutate(clust = case_when(id == "CD4 T cells" ~ "CD4 T cells (3617)",
                           id == "CD8 T cells" ~ "CD8 T cells (657)",
                           id == "NK cells" ~ "NK cells (673)",
                           id == "DCs" ~ "DCs (83)",
                           id == "B cells" ~ "B cells (23)"))

s1_dotplot_data$clust <- factor(s1_dotplot_data$clust, levels = c("CD4 T cells (3617)", "CD8 T cells (657)", "NK cells (673)", "DCs (83)", "B cells (23)"))

# Create dotplot
s1_fancy_dotplot <- s1_dotplot_data %>% 
  dplyr::filter(features.plot %in% adt_markers) %>%
  dplyr::filter(pct.exp > 0) %>% # drop marker-clusters with 0 % cells expressing any amount of the marker
  ggplot(aes(x = reorder(clust, desc(clust)), y = features.plot, color = avg.exp.scaled, size = pct.exp)) +
  geom_point() +
  coord_flip() +
  scale_color_gradient2(low = "#196293", mid = "white", high = "#9D1213") +
  theme_bw() +
  theme(axis.line = element_blank(),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(), 
        axis.text.x = element_text(color = "black", size = 9, angle = 90, hjust = 1, vjust = 0.5),
        axis.text.y = element_text(color = "black", size = 9),
        plot.title = element_text(hjust = 0.5, size = 12),
        panel.spacing.x = unit(0, "cm"),
        panel.border = element_blank(),
        axis.line.y.left = element_line(color = "black"),
        axis.line.x.bottom = element_line(color = "black"),
        legend.position="none",
        panel.grid = element_blank()) +
  labs(title = "S1",
       y = "Features") 

# Extract legend
s1_dotplot_legend <- as_ggplot(get_legend(s1_fancy_dotplot +
                                            theme(legend.position="bottom",
                                                  legend.box = "vertical", 
                                                  legend.title = element_text(size = 9),
                                                  legend.text = element_text(size = 9)) +
                                            labs(color = "Average Expression",
                                                 size = "Percent Expressed") +
                                            guides(color = guide_colorbar(order = 1),
                                                   size = guide_legend(order = 0))))
s1_fancy_dotplot
s1_dotplot_legend
```

# Filter out non-T cells

Let's subset all of the T cells.

```{r, fig.width=9, fig.height=7}
# How many cells do we have before filtering out non-T cells?
table(s1$MULTI_ID_global)

s1_sub <- subset(s1, subset = adt_clusters %in%
                               c("CD4 T cells", "CD8 T cells"))

# How many cells do we have after filtering out non-T cells?
table(s1_sub$MULTI_ID_global)
```

# Save output

```{r save Seurat object}
if(save_output) {
  saveRDS(s1, file = here::here("processed_data/20220323_s1_with_adt_labels.rds"))
  saveRDS(s1_sub, file = here::here("processed_data/20220323_s1_tcells_only.rds"))
  
  cairo_pdf(file=here::here("out/S1_ADT_dotplot.pdf"), width=3.5, height=2,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(s1_fancy_dotplot)
  dev.off()
  
  cairo_pdf(file=here::here("out/S1_ADT_dotplot_legend.pdf"), width=4, height=2,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(s1_dotplot_legend)
  dev.off()
}
```
