==> scripts/06_HAARVIVAC_10X_S1_WNN_Cluster_Annotation.Rmd <==
---
title: "HAARVIVAC 10X S1 WNN Cluster Annotation"
author: "Jolie Phan"
output: 
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: 4
date: "version `r format(Sys.time(), '%B %d, %Y')`"
editor_options: 
  chunk_output_type: console
---

The goal of this analysis is to annotate the Spike 1 WNN-derived clusters.

What lineage-defining TFs/cytokines/chemokines are differentially expressed among the WNN clusters? 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries

```{r, message = FALSE}
library(readxl)
library(Seurat)
library(dplyr)
library(ggplot2)
library(ggbeeswarm)
library(patchwork)
library(aplot)
library(ComplexHeatmap)
library(RColorBrewer)
library(scales)
```

# Load data

```{r}
rerun_de <- FALSE
save_output <- TRUE
rescale <- TRUE

if(save_output) {
  s1 <- readRDS(file = here::here("processed_data/20220323_s1_tcells_with_wnn.rds"))
} else {
  s1 <- readRDS(file = here::here("processed_data/20220413_s1_tcells_with_wnn_names.rds"))
}

genes4heatmap <- read.csv(here::here("data/Genes4HeatMap - Fig3HeatMap.csv"))

if(!dir.exists(here::here("out/annotation_plots"))) {
    cat(sprintf("Creating folder %s\n", here::here("out/annotation_plots")))
    dir.create(here::here("out/annotation_plots"), recursive = T)
}

if(!dir.exists(here::here("out/UMAP"))) {
    cat(sprintf("Creating folder %s\n", here::here("out/UMAP")))
    dir.create(here::here("out/UMAP"), recursive = T)
}
```

# Unnamed ADT violin plots
Look at the surface protein expression levels for each cluster.

First, list the ADT markers used in the CITE-seq panel. 

```{r}
lin_adt_feat <- c("CD3", "CD4", "CD8a", "GD", "TRAV1-2")
lin2_adt_feat <- c("CD14", "CD56", "CD19", "CD123", "CD15", "CD20", "CD163")
lin3_adt_feat <- c("CD16", "CD103", "CD161", "CD169", "CD11b", "CD11c")
act_adt_feat <- c("CD95", "CD25", "CD38", "CD26", "HLA-DR")
mem_adt_feat <- c("CD62L", "CD127", "CCR7", "CD45RA", "CD28")
ck_adt_feat <- c("CXCR3", "CXCR5", "CCR4", "CCR5", "CCR6")
fxn_adt_feat <- c("PD-L1", "CD86")
```

Generate violin plots of the background-corrected, normalized surface protein expression levels. 

Clusters 0-7, 11, 13, and 14 are CD4 T cells.
Clusters 8-10 and 12 are CD8 T cells. 
Cluster 10 is CD8 MAIT.
Maybe Cluster 14 is GD? nope

```{r, fig.width=12, fig.height=10}
Idents(s1) <- "wsnn_res.1"

lin_plot <- VlnPlot(s1, assay = "ADT_bc", features = lin_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

lin2_plot <- VlnPlot(s1, assay = "ADT_bc", features = lin2_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

lin3_plot <- VlnPlot(s1, assay = "ADT_bc", features = lin3_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

act_plot <- VlnPlot(s1, assay = "ADT_bc", features = act_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

mem_plot <- VlnPlot(s1, assay = "ADT_bc", features = mem_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

ck_plot <- VlnPlot(s1, assay = "ADT_bc", features = ck_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

fxn_plot <- VlnPlot(s1, assay = "ADT_bc", features = fxn_adt_feat, pt.size = 0.5) +
  plot_annotation(title = "Spike 1") & 
  theme(plot.title = element_text(hjust = 0.5))

if(save_output) {
  cairo_pdf(file=here::here("out/20220331_S1_WNN_ADT_Violin_Plots.pdf"), width=12, height=10, onefile = TRUE)
  print(lin_plot)
  print(lin2_plot)
  print(lin3_plot)
  print(act_plot)
  print(mem_plot)
  print(ck_plot)
  print(fxn_plot)
  dev.off()
}
```

# Unnamed ADT dot plot

```{r}
Idents(s1) <- "wsnn_res.1"
adt_markers <- c("CD3", "CD4", "CD8a", "GD", "TRAV1-2", "CD56", "CD95", "CD62L", "CD127", "CCR7", "CD45RA", "CXCR3", "CXCR5", "CCR4", "CCR5", "CCR6", "CD28", "CD86", "CD103", "CD25", "CD38", "CD26", "HLA-DR", "PD-L1", "CD161", "CD14", "CD123", "CD15", "CD163", "CD16", "CD11b", "CD11c", "CD169", "CD19", "CD20")

# Make dot plot
s1_dotplot <- DotPlot(s1, features = adt_markers, assay = "ADT_bc") + 
  RotatedAxis() +
  ggtitle("Spike 1 Bg-corr ADT Expression") +
  theme(plot.title = element_text(hjust = 0.5))
s1_dotplot 
```

# DE Testing 

```{r}
DefaultAssay(s1) <- "RNA"
Idents(s1) <- "wsnn_res.1"

if(rerun_de) {
  s1_markers <- FindAllMarkers(s1, only.pos = TRUE)
} else {
  s1_markers <- readRDS(here::here("out/20220331_S1_all_DE_genes.rds"))
}

s1_signif_markers <- s1_markers %>%
  dplyr::filter(p_val_adj < 0.05)

if(save_output) {
  saveRDS(s1_markers, here::here("out/20220331_S1_all_DE_genes.rds"))
  write.csv(s1_markers, here::here("out/20220331_S1_all_DE_genes.csv"), row.names = FALSE)
}

# Grab the top 5
s1_markers %>%
    group_by(cluster) %>%
    top_n(n = 5, wt = avg_log2FC) -> s1_top5
```

# Assign cell type names

```{r}
Idents(s1) <- "wsnn_res.1"

# Assign names to clusters
new_cluster_ids <- c("Activated Cytotoxic CD4", "Activated Th1 CD4", "CCR6+ IFIT+ CD4", "Naive-like CD4", "TREG IL22 CD4", "CCR6+CCR4+ CD4", "CCR6-CCR4- CXCR4+ CD4", "CCR5+CD26+CD161+ Cytotoxic CD4", "TEMRA Cytotoxic CD8", "Activated Cytotoxic CD8", "MAIT CD8", "TRM CD4", "Naive-like CD8", "CD26+CD161+ STAT4+CCL4+ CD4", "GD")
names(new_cluster_ids) <- levels(s1)
s1 <- RenameIdents(s1, new_cluster_ids)
s1[["wnn_clusters"]] <- Idents(s1)
```

# Named ADT dot plot

```{r out.extra='style="max-width:none; width:100vw; margin-left:calc(50% - 50vw);"', fig.width = 13, fig.height = 6}
Idents(s1) <- "wnn_clusters"

# List markers to include for plotting
adt_subset <- c("CD3", "CD4", "CD8a", "GD", "TRAV1-2", "CD56")
adt_mem <- c("CD95", "CD62L", "CD127", "CCR7", "CD45RA")
adt_homing <- c("CXCR3", "CXCR5", "CCR4", "CCR5", "CCR6")
adt_act <- c("CD28", "CD86", "CD103", "CD25", "CD38", "CD26", "HLA-DR", "PD-L1", "CD161", "CD11b", "CD11c")
adt_tcell_markers <- c(adt_subset, adt_mem, adt_homing, adt_act)

# Make a Seurat dot plot just to easily pull the necessary data
s1_dotplot_named <- DotPlot(s1, features = adt_tcell_markers, assay ="ADT_bc") + 
  RotatedAxis() +
  ggtitle("Spike 1 Bg-corr ADT Expression") +
  theme(plot.title = element_text(hjust = 0.5))

# Pull the necessary data from the Seurat-generated dotplot 
s1_dotplot_data <- s1_dotplot_named$data

markers <- s1_dotplot_data$features.plot %>%
  unique()

# Add ADT categories to data
s1_dotplot_data <- s1_dotplot_data %>%
  mutate(category = case_when(features.plot %in% adt_subset ~ "T Cell Subset",
                              features.plot %in% adt_mem ~ "Memory",
                              features.plot %in% adt_homing ~ "Homing",
                              features.plot %in% adt_act ~ "Activation"))
adt_cat_names <- c("T Cell Subset", "Memory", "Homing", "Activation")
s1_dotplot_data$category <- factor(s1_dotplot_data$category, levels = adt_cat_names)

# Add cluster numbers to cluster names
s1_dotplot_data <- s1_dotplot_data %>%
  mutate(clust = case_when(id == "Activated Cytotoxic CD4" ~ "0: Activated Cytotoxic CD4",
                           id == "Activated Th1 CD4" ~ "1: Activated Th1 CD4",
                           id == "CCR6+ IFIT+ CD4" ~ "2: CCR6+ IFIT+ CD4",
                           id == "Naive-like CD4" ~ "3: Naive-like CD4",
                           id == "TREG IL22 CD4" ~ "4: TREG IL22 CD4",
                           id == "CCR6+CCR4+ CD4" ~ "5: CCR6+CCR4+ CD4",
                           id == "CCR6-CCR4- CXCR4+ CD4" ~ "6: CCR6-CCR4- CXCR4+ CD4",
                           id == "CCR5+CD26+CD161+ Cytotoxic CD4" ~ "7: CCR5+CD26+CD161+ Cytotoxic CD4",
                           id == "TEMRA Cytotoxic CD8" ~ "8: TEMRA Cytotoxic CD8",
                           id == "Activated Cytotoxic CD8" ~ "9: Activated Cytotoxic CD8",
                           id == "MAIT CD8" ~ "10: MAIT CD8",
                           id == "TRM CD4" ~ "11: TRM CD4",
                           id == "Naive-like CD8" ~ "12: Naive-like CD8",
                           id == "CD26+CD161+ STAT4+CCL4+ CD4" ~ "13: CD26+CD161+ STAT4+CCL4+ CD4",
                           id == "GD" ~ "14: GD"))
s1_dotplot_data$clust <- factor(s1_dotplot_data$clust, levels = c("0: Activated Cytotoxic CD4",
                                                                  "1: Activated Th1 CD4",
                                                                  "2: CCR6+ IFIT+ CD4",
                                                                  "3: Naive-like CD4",
                                                                  "4: TREG IL22 CD4",
                                                                  "5: CCR6+CCR4+ CD4",
                                                                  "6: CCR6-CCR4- CXCR4+ CD4",
                                                                  "7: CCR5+CD26+CD161+ Cytotoxic CD4",
                                                                  "8: TEMRA Cytotoxic CD8",
                                                                  "9: Activated Cytotoxic CD8",
                                                                  "10: MAIT CD8",
                                                                  "11: TRM CD4",
                                                                  "12: Naive-like CD8",
                                                                  "13: CD26+CD161+ STAT4+CCL4+ CD4",
                                                                  "14: GD"))

# Create dotplot
s1_fancy_dotplot <- s1_dotplot_data %>% 
  dplyr::filter(features.plot %in% markers) %>%
  dplyr::filter(pct.exp > 0) %>% # drop marker-clusters with 0 % cells expressing any amount of the marker
  ggplot(aes(x = reorder(clust, desc(clust)), y = features.plot, color = avg.exp.scaled, size = pct.exp)) +
  facet_grid(. ~ category, space = "free_x", scales = "free_x", switch = "x") +
  geom_point() +
  coord_flip() +
  scale_color_gradient2(low = "#196293", mid = "white", high = "#9D1213") +
  theme_bw() +
  theme(axis.line = element_blank(),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(), 
        axis.text.x = element_text(color = "black", size = 7, angle = 90, hjust = 1, vjust = 0.5),
        axis.text.y = element_text(color = "black", size = 7),
        plot.title = element_text(hjust = 0.5, size = 12),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 7),
        strip.placement = "outside",
        strip.background = element_rect(fill = NA, color = "black"),
        strip.text = element_text(size = 7), 
        panel.spacing.x = unit(0, "cm"),
        panel.border = element_blank(),
        axis.line.y.left = element_line(color = "black"),
        axis.line.x.bottom = element_line(color = "black"),
        legend.position="bottom",
        panel.grid = element_blank()) +
  labs(title = "S1",
       color = "Average Expression ",
       size = "Percent Expressed") +
  guides(color = guide_colorbar(order = 1),
         size = guide_legend(order = 0))

s1_fancy_dotplot
```

# Named GEX Heat map

```{r, fig.width=7, fig.height=3}
Idents(s1) <- "wsnn_res.1"
DefaultAssay(s1) <- "RNA"

# Rescale RNA data to include all genes in heat map
if(rescale) {
  s1 <- ScaleData(object = s1, vars.to.regress = "percent.mt", features = rownames(s1), block.size = 2000)
}

# Find average expression
s1_avg <- AverageExpression(s1, return.seurat = TRUE, assays = "RNA", slot = "scale.data")

# Grab data for plotting
mat <- s1_avg[["RNA"]]@scale.data[genes4heatmap$Gene,] %>%
  as.matrix() %>%
  t()

# cluster_anno <- s1_avg@active.ident
col_fun <- circlize::colorRamp2(c(-2, 0, 2), c("#196293", "white", "#9D1213"))

# Plot heat map
s1_heatmap <- Heatmap(mat, name = "Average Expression",
                      cluster_columns = FALSE,
                      cluster_rows = FALSE,
                      show_column_dend = FALSE,
                      show_row_dend = FALSE,
                      row_names_side = "left",
                      column_names_gp = grid::gpar(fontsize = 8),
                      row_names_gp = grid::gpar(fontsize = 8),
                      column_title = "S1",
                      col = col_fun
                      )

s1_heatmap
```

# UMAP

```{r, fig.height=3, fig.width=4}
Idents(s1) <- "wsnn_res.1"

num_cols <- 15
umap_colors <- colorRampPalette(brewer.pal(12, "Set3"))(num_cols)

s1_umap <- DimPlot(s1, reduction = "wnn.umap", label = TRUE, repel = TRUE, label.size = 3, pt.size = 0.3, cols = umap_colors) + 
  ggtitle("S1") +
  theme_bw() +
  theme(plot.title = element_text(size = 12, face = "plain", hjust = 0.5),
        axis.text.x = element_text(size = 8, color = "black"),
        axis.text.y = element_text(size = 8, color = "black"),
        axis.title.x = element_text(size = 8),
        axis.title.y = element_text(size = 8),
        axis.line.y.left = element_line(color = "black"),
        axis.line.x.bottom = element_line(color = "black"),
        axis.line = element_blank(),
        panel.border = element_blank(),
        panel.grid = element_blank(),
        legend.position = "none") 
s1_umap
```

# Save Seurat Object

```{r}
if(save_output) {
  saveRDS(s1, file = here::here("processed_data/20220413_s1_tcells_with_wnn_names.rds"))
}
```

# Save plots

```{r}
if(save_output) {
  cairo_pdf(file=here::here("out/annotation_plots/S1_GEX_heatmap.pdf"), width=5.6, height=2.55,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(s1_heatmap)
  dev.off()
  
  cairo_pdf(file=here::here("out/annotation_plots/S1_ADT_dotplot.pdf"), width=8.5, height=5,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(s1_fancy_dotplot)
  dev.off()
  
  cairo_pdf(file=here::here("out/UMAP/S1_WNN_UMAP.pdf"), width=3, height=3,
        onefile = TRUE, bg = "transparent", family = "Arial") # default unit is inches, default font is Helvetica.
  print(s1_umap)
  dev.off()
}
```
